[FD-208] feat: 일정 추가와 수정을 Action으로 합치고 ActionType으로 구분
